pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-demo"
version = "0.0.1"
description = "A demo API using FastAPI"
authors = [{ name = "Bassem Karoui", email = "bassem.karoui1@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
keywords = [
    "fastapi-demo",
    "python",
]
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi[standard]>=0.115.12",
    "slowapi>=0.1.9",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
    "orjson>=3.10.18",
    "fastapi-users[oauth,redis]>=14.0.1",
    "fastapi-users-db-sqlmodel>=0.3.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.21.0",
    "fastapi-mail>=1.5.0",
    "alembic>=1.16.1",
    "limits>=5.2.0",
    "aiocache>=0.12.3",
    "asgiref>=3.8.1",
    "celery[redis]>=5.5.3",
    "flower>=2.0.1",
    "openfga-sdk>=0.9.5",
    "alembic-postgresql-enum>=1.8.0",
    "librabbitmq",
    "fastapi-pagination[sqlmodel]",
    "python-ulid[pydantic]>=3.0.0",
]

[project.urls]
Homepage = "https://bassemkaroui.github.io/fastapi-demo"
Repository = "https://github.com/bassemkaroui/fastapi-demo"
Documentation = "https://bassemkaroui.github.io/fastapi-demo"

# [project.scripts]
# fastapi-demo-prestart = "fastapi_demo.cli.prestart:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pre-commit>=4.2.0",
    "tox-uv>=1.25.0",
    "mypy>=1.15.0",
    "ruff>=0.11.5",
    "commitizen>=4.6.0",
    "cz-conventional-gitmoji>=0.7.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.11",
    "mkdocstrings[python]>=0.29.1",
    "markdown-callouts>=0.4.0",
    "mkdocs-minify-plugin>=0.8.0",
    "griffe-typingdoc>=0.2.8",
    "griffe>=1.7.2",
    "ipython>=8.35.0",
    "duty>=1.6.0",
    "pytest-asyncio>=0.26.0",
    "debugpy>=1.8.14",
]

[tool.commitizen]  # https://commitizen-tools.github.io/commitizen/config/
name = "cz_gitmoji"
version_files = ["src/fastapi_demo/__init__.py", "compose.yaml"]
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
changelog_merge_prerelease = true
gpg_sign = true

[tool.mypy]
files = ["src"]
exclude = ["src/dataup_pipelines/alembic"]
check_untyped_defs = true
disallow_untyped_defs = true
# disallow_any_unimported = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["fastapi_demo.api.routes.*", "fastapi_demo.api.main"]
disallow_untyped_defs = false

[tool.pyright]
include = ["src"]
pythonVersion = "py312"
typeCheckingMode = "off"
useLibraryCodeForTypes = true
reportUndefinedVariable = "none"

[tool.ruff]
target-version = "py310"
line-length = 100
fix = true
exclude = ["alembic"]

[tool.ruff.lint]
preview = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # flake8-future-annotations
    "FA",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
    # NumPy-specific rules
    "NPY",
    # pandas-vet
    "PD",
    # pep8-naming
    "N",
    # flake8-pytest-style
    "PT",
    # flake8-pie
    "PIE",
    # flake8-tidy-imports
    "TID",
    # flake8-implicit-str-concat
    "ISC",
    # Perflint
    "PERF",
    # flynt
    "FLY",
    # unconventional-import-alias
    "ICN001",
    # pylint
    "PL",
    # flake8-unused-arguments
    "ARG",
    # FastAPI
    "FAST",
]
ignore = [
    # LineTooLong
    "E501",
    # indentation-with-invalid-multiple
    "E111",
    # indentation-with-invalid-multiple-comment
    "E114",
    # over-indented
    "E117",
    # tab-indentation
    "W191",
    # too-many-public-methods
    "PLR0904",
    # Too many returns
    "PLR0911",
    # Too many branches
    "PLR0912",
    # Too many arguments to function call
    "PLR0913",
    # too-many-locals
    "PLR0914",
    # Too many statements
    "PLR0915",
    # too-many-boolean-expressions
    "PLR0916",
    # Too many positional arguments
    "PLR0917",
    # too-many-nested-blocks
    "PLR1702",
    # literal-membership
    "PLR6201",
    # Private name import
    "PLC2701",
    # compare-to-empty-string
    "PLC1901",
    # import-outside-toplevel
    "PLC0415",
    # unnecessary-dunder-call
    "PLC2801",
    # Global statements are discouraged
    "PLW0603",
    # unspecified-encoding
    "PLW1514",
    # raise-vanilla-args
    "TRY003",

]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"src/fastapi_demo/api/exception_handlers.py" = ["RUF029", "ARG001"]
"src/fastapi_demo/api/dependencies.py" = ["RUF029", "ARG001"]
"src/fastapi_demo/core/config.py" = ["N802"]

[tool.ruff.format]
preview = true
docstring-code-format = true

# [tool.ruff.lint.isort]
# # Useful when using Python < 3.10
# required-imports = ["from __future__ import annotations"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]

[tool.coverage.report]
fail_under = 100
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug",
    "if __name__ == .__main__.:",
]
omit = ["*/__init__.py", "*/fastapi_demo/testing/*"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.uv.sources]
librabbitmq = { git = "https://github.com/celery/librabbitmq.git" }
fastapi-pagination = { git = "https://github.com/uriyyo/fastapi-pagination.git" }
